requirements.txt
openai==2.7.1
python-dotenv==1.2.1
pandas==2.2.2
scikit-learn==1.5.1
matplotlib==3.9.2
seaborn==0.13.2
tqdm==4.66.4
datasets==4.4.1
langchain==1.0.7
langchain-openai==1.0.3
langgraph==1.0.3